backend/requirements.txt
# Core
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1

# Database
motor>=3.4.0
pymongo>=4.7.0

# AI / LLM
openai>=1.30.0
google-generativeai>=0.5.0
anthropic>=0.26.0

# Audio transcription
faster-whisper>=1.0.0
torch>=2.2.0

# HTTP (async)
httpx>=0.27.0

# Utilities
python-multipart>=0.0.9

# Optional: local translation
# transformers>=4.40.0
# tqdm>=4.66.0